requirements.txt
asgiref==3.8.1
astroid==3.3.5
dill==0.3.9
dj-database-url==2.2.0
Django>=4.2
djangorestframework==3.15.2
isort==5.13.2
mccabe==0.7.0
platformdirs==4.3.6
psycopg2==2.9.9
pylint==3.3.1
python-dotenv==1.0.1
sqlparse==0.5.1
tomlkit==0.13.2
typing_extensions==4.12.2
